third_party/blink/web_tests/editing/editability/empty-document-stylewithcss-expected.txt

CONSOLE WARNING: document.execCommand() doesn't work with an invalid HTML structure. It is corrected automatically.
CONSOLE WARNING: document.execCommand() doesn't work with an invalid HTML structure. It is corrected automatically.
This test ensures WebKit executes StyleWithCSS properly even in an empty document.
First negation:PASS
Second negation:PASS
